Initial commit
This commit creates two packages: `approaches` and `fixedvalue`.
`approaches` demonstrates four ways to test the same function.
`fixedvalue` is a trivial gin app with a single HTTP endpoint that
can be set up to run the integration tests in the `approaches`
folder.

I thought I was going to do this in two commits with the unit tests
coming first and the integration setup second, but adding the
`fixedvalue` package forced me to put all the approach folders in a
subdirectory, and committing the subdirectory without the rest of
the integration setup didn't make sense out of context.